Support Null Return for Enricher
Currently, if the enricher flow returns null, the enricher returns null.

See: https://jira.spring.io/browse/INT-3465

In this work around, we invoke the "enricher" flow from a header
enricher and store the result in a header.

We then use an enricher to extract the required field from the
flow (if not null).

Finally, we use a header filter to remove the temporary header.